+++ src/test/java/school/redrover/ViewTest.java
@@ -1,14 +1,21 @@
 package school.redrover;
 
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.*;
+import org.openqa.selenium.interactions.Actions;
 import org.testng.Assert;
-import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.runner.BaseTest;
-import org.openqa.selenium.Alert;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class ViewTest extends BaseTest {
+
+    private static final String JOB_NAME = "FreestyleProject-1";
+    private static final String JOB_NAME_1 = "FreestyleProject-2";
+    private static final String JOB_NAME_2 = "FreestyleProject-3";
+    private static final String VIEW_NAME = "ListView-1";
+    private static final String VIEW_NAME_1 = "ListView-new";
 
     private void goHome() {
         getDriver().findElement(By.id("jenkins-home-link")).click();
@@ -29,6 +36,40 @@ public class ViewTest extends BaseTest {
         getDriver().findElement(By.cssSelector("input[name='name']")).sendKeys(viewName);
         getDriver().findElement(By.xpath("//label[@for='hudson.model.ListView']")).click();
         getDriver().findElement(By.cssSelector("button[id='ok']")).click();
+    }
+
+    private void createListViewWithoutAssociatedJob(String newListViewName) {
+        goHome();
+        getDriver().findElement(By.xpath("//a[@href = '/newView']")).click();
+        getDriver().findElement(By.xpath("//input[@name = 'name']")).sendKeys(newListViewName);
+        getDriver().findElement(By.xpath("//label[@for = 'hudson.model.ListView']")).click();
+        getDriver().findElement(By.xpath("//button[@id = 'ok']")).click();
+    }
+
+    private void createListViewWithAssociatedJob(String newListViewName) {
+        goHome();
+        getDriver().findElement(By.xpath("//a[@href = '/newView']")).click();
+        getDriver().findElement(By.xpath("//input[@name = 'name']")).sendKeys(newListViewName);
+        getDriver().findElement(By.xpath("//label[@for = 'hudson.model.ListView']")).click();
+        getDriver().findElement(By.xpath("//button[@id = 'ok']")).click();
+        getDriver().findElement(By.xpath("//div[@class = 'listview-jobs']/span/span[1]")).click();
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+    }
+
+    private void addNewDescriptionForTheView(String listViewName, String newDescriptionForTheView) {
+        goHome();
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + listViewName + "/']")).click();
+        getDriver().findElement(By.xpath("//a[@id = 'description-link']")).click();
+        getDriver().findElement(By.xpath("//textarea[@name = 'description']")).sendKeys(newDescriptionForTheView);
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+    }
+
+    private void associateJobToTheView(String listViewName, String jobName) {
+        goHome();
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + listViewName + "/']")).click();
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + listViewName + "/configure']")).click();
+        getDriver().findElement(By.xpath("//label[@title = '" + jobName + "']")).click();
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
     }
 
     @Test
@@ -204,5 +245,297 @@ public class ViewTest extends BaseTest {
 
         Assert.assertTrue(text.contains(nameView));
 
+    }
+
+    @Test
+    public void testAddingDescriptionForTheView() {
+        final String newFreeStyleProjectName = "FreeStyleTestProject";
+        final String newListViewName = "ListViewTest";
+        final String newDescriptionForTheView = "Test description for the List View";
+
+        createNewFreestyleProject(newFreeStyleProjectName);
+        createListViewWithoutAssociatedJob(newListViewName);
+        goHome();
+
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/']")).click();
+        getDriver().findElement(By.xpath("//a[@id = 'description-link']")).click();
+        getDriver().findElement(By.xpath("//textarea[@name = 'description']")).sendKeys(newDescriptionForTheView);
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+
+        Assert.assertEquals(
+                getDriver().findElement(By.xpath("//div[@id = 'description']/div[1]")).getText(),
+                newDescriptionForTheView);
+    }
+
+    @Test
+    public void testEditingDescriptionForTheView() {
+        final String newFreeStyleProjectName = "FreeStyleTestProject";
+        final String newListViewName = "ListViewTest";
+        final String newDescriptionForTheView = "Test description for the List View";
+        final String editedDescriptionForTheView = "New Test description for the List View instead of the previous one";
+
+        createNewFreestyleProject(newFreeStyleProjectName);
+        createListViewWithoutAssociatedJob(newListViewName);
+        addNewDescriptionForTheView(newListViewName, newDescriptionForTheView);
+        goHome();
+
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/']")).click();
+        getDriver().findElement(By.xpath("//a[@id = 'description-link']")).click();
+        getDriver().findElement(By.xpath("//textarea[@name = 'description']")).clear();
+        getDriver().findElement(By.xpath("//textarea[@name = 'description']")).sendKeys(editedDescriptionForTheView);
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+
+        Assert.assertEquals(
+                getDriver().findElement(By.xpath("//div[@id = 'description']/div[1]")).getText(),
+                editedDescriptionForTheView);
+    }
+
+    @Test
+    public void testDeletingDescriptionForTheView() {
+        final String newFreeStyleProjectName = "FreeStyleTestProject";
+        final String newListViewName = "ListViewTest";
+        final String newDescriptionForTheView = "Test description for the List View";
+
+        createNewFreestyleProject(newFreeStyleProjectName);
+        createListViewWithoutAssociatedJob(newListViewName);
+        addNewDescriptionForTheView(newListViewName, newDescriptionForTheView);
+        goHome();
+
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/']")).click();
+        getDriver().findElement(By.xpath("//a[@id = 'description-link']")).click();
+        getDriver().findElement(By.xpath("//textarea[@name = 'description']")).clear();
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+
+        Assert.assertEquals(
+                getDriver().findElement(By.xpath("//div[@id = 'description']/div[1]")).getText(),
+                "");
+    }
+
+    @Test
+    public void testNoJobsShownForTheViewWithoutAssociatedJob() {
+        final String newFreeStyleProjectName = "FreeStyleTestProject";
+        final String newListViewName = "ListViewTest";
+        final String noAssociatedJobsForTheViewMessage = "This view has no jobs associated with it. " +
+                "You can either add some existing jobs to this view or create a new job in this view.";
+
+        createNewFreestyleProject(newFreeStyleProjectName);
+        createListViewWithoutAssociatedJob(newListViewName);
+        goHome();
+
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/']")).click();
+
+        Assert.assertTrue(
+                getDriver().findElement(By.xpath("//div[@id = 'main-panel']")).getText().
+                        contains(noAssociatedJobsForTheViewMessage));
+    }
+
+    @Test
+    public void testProjectCouldBeAddedToTheView() {
+        final String newFreeStyleProjectName = "FreeStyleTestProject";
+        final String newListViewName = "ListViewTest";
+
+        createNewFreestyleProject(newFreeStyleProjectName);
+        createListViewWithoutAssociatedJob(newListViewName);
+        goHome();
+
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/']")).click();
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/configure']")).click();
+        getDriver().findElement(By.xpath("//label[@title = '" + newFreeStyleProjectName + "']")).click();
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+
+        Assert.assertEquals(
+                getDriver().findElement(By.xpath("//a[@class = 'jenkins-table__link model-link inside']")).getText(),
+                newFreeStyleProjectName);
+    }
+
+    @Test
+    public void testAssociatedJobIsShownOnTheViewDashboard() {
+        final String newFreeStyleProjectName = "FreeStyleTestProject";
+        final String newListViewName = "ListViewTest";
+
+        createNewFreestyleProject(newFreeStyleProjectName);
+        createListViewWithoutAssociatedJob(newListViewName);
+        associateJobToTheView(newListViewName, newFreeStyleProjectName);
+        goHome();
+
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/']")).click();
+
+        Assert.assertEquals(
+                getDriver().findElement(By.xpath("//a[@class = 'jenkins-table__link model-link inside']")).getText(),
+                newFreeStyleProjectName);
+    }
+
+    @Test
+    public void testAddingNewColumnToTheView() {
+        final String newFreeStyleProjectName = "FreeStyleTestProject";
+        final String newListViewName = "ListViewTest";
+        final String newColumnName = "Git Branches";
+
+        createNewFreestyleProject(newFreeStyleProjectName);
+        createListViewWithAssociatedJob(newListViewName);
+        goHome();
+
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/']")).click();
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/configure']")).click();
+        JavascriptExecutor scriptForScrolling = (JavascriptExecutor) getDriver();
+        scriptForScrolling.executeScript("window.scrollBy(0,926)");
+        getDriver().findElement(By.xpath("//button[@id = 'yui-gen3-button']")).click();
+
+        List<WebElement> newColumnOptions = getDriver().findElements(By.xpath("//a[@class = 'yuimenuitemlabel']"));
+        for (WebElement newColumnOption : newColumnOptions) {
+            if (newColumnOption.getText().contains(newColumnName)) {
+                newColumnOption.click();
+            }
+        }
+
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+
+        List<WebElement> dashboardColumnNames = getDriver().findElements(By.xpath("//table[@id = 'projectstatus']//th"));
+
+        Assert.assertEquals(
+                dashboardColumnNames.get(dashboardColumnNames.size()-1).getText(),
+                newColumnName);
+    }
+
+    @Test
+    public void testDeletingColumnFromTheView() {
+        final String newFreeStyleProjectName = "FreeStyleTestProject";
+        final String newListViewName = "ListViewTest";
+        final String deletedColumnName = "Last Duration";
+
+        createNewFreestyleProject(newFreeStyleProjectName);
+        createListViewWithAssociatedJob(newListViewName);
+        goHome();
+
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/']")).click();
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/configure']")).click();
+        JavascriptExecutor scriptForScrolling = (JavascriptExecutor) getDriver();
+        scriptForScrolling.executeScript("window.scrollBy(0,926)");
+        getDriver().findElement(By.xpath(
+                "//div[contains(text(), '" + deletedColumnName + "')]/button")).click();
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+        List<WebElement> dashboardColumnNamesAfterColumnDeletion = getDriver().findElements(By.xpath(
+                "//table[@id = 'projectstatus']//th"));
+
+        Assert.assertFalse(dashboardColumnNamesAfterColumnDeletion.contains(deletedColumnName));
+    }
+
+    @Test
+    public void testReorderColumnsForTheView() {
+        final String newFreeStyleProjectName = "FreeStyleTestProject";
+        final String newListViewName = "ListViewTest";
+        final String reorderedColumnName = "Name";
+
+        createNewFreestyleProject(newFreeStyleProjectName);
+        createListViewWithAssociatedJob(newListViewName);
+        goHome();
+
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/']")).click();
+        getDriver().findElement(By.xpath("//a[@href = '/view/" + newListViewName + "/configure']")).click();
+        JavascriptExecutor scriptForScrolling = (JavascriptExecutor) getDriver();
+        scriptForScrolling.executeScript("window.scrollBy(0,926)");
+        WebElement columnToReorder = getDriver().findElement(By.xpath(
+                "//div[contains(text(), '" + reorderedColumnName + "')]/div[@class = 'dd-handle']"));
+        WebElement placeForTheReorderedColumn = getDriver().findElement(By.xpath(
+                "//div[@class = 'repeated-chunk__header'][1]"));
+        Actions actions = new Actions(getDriver());
+        actions.clickAndHold(columnToReorder)
+                .moveToElement(placeForTheReorderedColumn)
+                .release(placeForTheReorderedColumn)
+                .build()
+                .perform();
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+        List<WebElement> dashboardColumnNamesAfterColumnReorder = getDriver().findElements(By.xpath(
+                "//table[@id = 'projectstatus']//th"));
+
+        Assert.assertTrue(dashboardColumnNamesAfterColumnReorder.get(0).getText().contains(reorderedColumnName));
+    }
+
+    @Test
+    public void testCreateNewListViewWithoutJobs() {
+        createNewFreestyleProject(JOB_NAME);
+        goHome();
+        createListViewWithoutAssociatedJob(VIEW_NAME);
+        goHome();
+
+        List<WebElement> listOfViews = getDriver().findElements(By.xpath("//div[@class='tabBar']/div"));
+        List<String> actualViewsNames = new ArrayList<>();
+        for (WebElement el : listOfViews) {
+            actualViewsNames.add(el.getText());
+        }
+
+        Assert.assertTrue(actualViewsNames.contains(VIEW_NAME));
+    }
+
+    @Test(dependsOnMethods = "testCreateNewListViewWithoutJobs")
+    public void testRenameListView() {
+        getDriver().findElement(By.xpath("//div[@class='tabBar']/div/a[@href='/view/" + VIEW_NAME + "/']"))
+                .click();
+        getDriver().findElement(By.xpath("//a[@href='/view/" + VIEW_NAME + "/configure']")).click();
+        getDriver().findElement(By.xpath("//input[@name='name']")).clear();
+        getDriver().findElement(By.xpath("//input[@name='name']")).sendKeys(VIEW_NAME_1);
+        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
+
+        Assert.assertEquals(getDriver().findElement(By.xpath("//div[@class = 'tab active']/a"))
+                        .getText(),
+                VIEW_NAME_1);
+    }
+
+    @Test(dependsOnMethods = "testRenameListView")
+    public void testAddJobToListViewWithoutJobs_fromMainSectionLink() {
+        getDriver().findElement(By.xpath("//div[@class='tabBar']/div/a[@href='/view/" + VIEW_NAME_1 + "/']"))
+                .click();
+
+        Assert.assertTrue(getDriver().findElement(By.id("main-panel")).getText()
+                .contains("This view has no jobs associated with it. You can either add some existing jobs to this " +
+                        "view or create a new job in this view."));
+
+        getDriver().findElement(By.xpath("//a[@href='configure']")).click();
+        JavascriptExecutor js = (JavascriptExecutor) getDriver();
+        js.executeScript("window.scrollBy(0,600)");
+        getDriver().findElement(By.xpath("//label[@title='" + JOB_NAME + "']")).click();
+        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
+
+        Assert.assertEquals(getDriver().findElement(By.xpath("//table[@id='projectstatus']/tbody/tr/td[3]"))
+                        .getText(),
+                JOB_NAME);
+    }
+
+    @Test(dependsOnMethods = "testAddJobToListViewWithoutJobs_fromMainSectionLink")
+    public void testAddAllJobsToView() {
+        createNewFreestyleProject(JOB_NAME_1);
+        goHome();
+        createNewFreestyleProject(JOB_NAME_2);
+        goHome();
+
+        getDriver().findElement(By.xpath("//div[@class='tabBar']/div/a[@href='/view/" + VIEW_NAME_1 + "/']"))
+                .click();
+        getDriver().findElement(By.xpath("//a[@href='/view/" + VIEW_NAME_1 + "/configure']")).click();
+
+        JavascriptExecutor js = (JavascriptExecutor) getDriver();
+        js.executeScript("window.scrollBy(0,600)");
+
+        getDriver().findElement(By.xpath("//label[@title='" + JOB_NAME + "']")).click();
+        List<WebElement> listOfJobs = getDriver().findElements(
+                By.xpath("//div[@class='listview-jobs']/span/span/label"));
+        for (WebElement job : listOfJobs) {
+            job.click();
+        }
+        getDriver().findElement(By.name("Submit")).click();
+
+        List<String> jobNames = List.of(JOB_NAME, JOB_NAME_1, JOB_NAME_2);
+        List<WebElement> dashboardItems = getDriver().findElements(
+                By.xpath("//table[@id='projectstatus']/tbody/tr/td"));
+        List<String> jobNamesDashboard = new ArrayList<>();
+        for (WebElement item : dashboardItems) {
+            if (item.getText().contains("FreestyleProject")) {
+                jobNamesDashboard.add(item.getText());
+            }
+        }
+
+        Assert.assertEquals(jobNamesDashboard.size(), jobNames.size());
+        for (int i = 0; i < jobNamesDashboard.size(); i++) {
+            Assert.assertEquals(jobNames.get(i), jobNamesDashboard.get(i));
+        }
     }
 }
